Move tasks to folder; revert eager change
